fix(daybot): fractional shares + blacklist chronic losers
POSITION SIZING BUG (root cause of poor performance):
- max(1, qty) forced minimum 1 share regardless of price
- AMD at $420 = 1 share = 42% of $1000 portfolio per trade
- TSLA at $405 = 1 share = 40% of portfolio per trade
- With 10% SL: $40+ loss per stop-hit, unsustainable at 49% WR
Fix: fractional shares via Alpaca — buy exactly 10% of portfolio worth
  $1000 * 10% / $420 = 0.238 shares AMD = $100 position (correct)

BLACKLIST chronic losers (from 133-trade history analysis):
- META: 4 trades, 25% WR, -$184.89 total (single biggest loser)
- MA:   6 trades, 17% WR,  -$70.33 total
- NVDA: 15 trades, 20% WR, -$37.72 total (also earnings-risk skipped)

WINNERS to focus on (100% WR): GOOGL +$85, BABA +$59, AMZN +$23
AMD remains tradeable: 16 trades, 50% WR, +$149 (fractional now)

+++ daybot/executor.py
@@ -37,25 +37,25 @@ class TradeExecutor:
                     time.sleep(wait)
         raise RuntimeError(f"All {retries} attempts failed")
 
-    def place_buy_order(self, symbol: str, qty: int) -> dict:
+    def place_buy_order(self, symbol: str, qty: float) -> dict:
         def _do():
             req = MarketOrderRequest(
                 symbol=symbol, qty=qty,
                 side=OrderSide.BUY, time_in_force=TimeInForce.DAY,
             )
             order = self._client.submit_order(req)
-            logging.info("BUY submitted: %s qty=%d id=%s", symbol, qty, order.id)
+            logging.info("BUY submitted: %s qty=%.4f id=%s", symbol, qty, order.id)
             return {"id": str(order.id), "symbol": symbol, "qty": qty, "side": "buy"}
         return self._with_retry(_do)
 
-    def place_sell_order(self, symbol: str, qty: int) -> dict:
+    def place_sell_order(self, symbol: str, qty: float) -> dict:
         def _do():
             req = MarketOrderRequest(
                 symbol=symbol, qty=qty,
                 side=OrderSide.SELL, time_in_force=TimeInForce.DAY,
             )
             order = self._client.submit_order(req)
-            logging.info("SELL submitted: %s qty=%d id=%s", symbol, qty, order.id)
+            logging.info("SELL submitted: %s qty=%.4f id=%s", symbol, qty, order.id)
             return {"id": str(order.id), "symbol": symbol, "qty": qty, "side": "sell"}
         return self._with_retry(_do)
